Refactor payment method routes into 1 route
