--- src/Compatibility/Storefront/Route/PaymentMethodRoute/Voucher/HideVoucherPaymentMethodRoute64.php
@@ -1,63 +1,0 @@
-<?php
-
-namespace Kiener\MolliePayments\Compatibility\Storefront\Route\PaymentMethodRoute\Voucher;
-
-use Kiener\MolliePayments\Service\Cart\Voucher\VoucherCartCollector;
-use Kiener\MolliePayments\Service\Cart\Voucher\VoucherService;
-use Kiener\MolliePayments\Service\Payment\Remover\PaymentMethodRemoverInterface;
-use Shopware\Core\Checkout\Cart\SalesChannel\CartService;
-use Shopware\Core\Checkout\Payment\SalesChannel\AbstractPaymentMethodRoute;
-use Shopware\Core\Checkout\Payment\SalesChannel\PaymentMethodRouteResponse;
-use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
-use Shopware\Core\System\SalesChannel\SalesChannelContext;
-use Symfony\Component\DependencyInjection\Container;
-use Symfony\Component\DependencyInjection\ContainerInterface;
-use Symfony\Component\HttpFoundation\Request;
-
-
-class HideVoucherPaymentMethodRoute64 extends AbstractPaymentMethodRoute
-{
-
-    /**
-     * @var AbstractPaymentMethodRoute
-     */
-    private $corePaymentMethodRoute;
-
-    /**
-     * @var PaymentMethodRemoverInterface
-     */
-    private $paymentMethodRemover;
-
-    /**
-     * @param AbstractPaymentMethodRoute    $corePaymentMethodRoute
-     * @param PaymentMethodRemoverInterface $paymentMethodRemover
-     */
-    public function __construct(AbstractPaymentMethodRoute $corePaymentMethodRoute, PaymentMethodRemoverInterface $paymentMethodRemover)
-    {
-        $this->corePaymentMethodRoute = $corePaymentMethodRoute;
-        $this->paymentMethodRemover = $paymentMethodRemover;
-    }
-
-
-    /**
-     * @return AbstractPaymentMethodRoute
-     */
-    public function getDecorated(): AbstractPaymentMethodRoute
-    {
-        return $this->corePaymentMethodRoute;
-    }
-
-    /**
-     * @param Request $request
-     * @param SalesChannelContext $context
-     * @param Criteria $criteria
-     * @return PaymentMethodRouteResponse
-     */
-    public function load(Request $request, SalesChannelContext $context, Criteria $criteria): PaymentMethodRouteResponse
-    {
-        $originalData = $this->corePaymentMethodRoute->load($request, $context, $criteria);
-
-        return $this->paymentMethodRemover->removePaymentMethods($originalData, $context);
-    }
-
-}
