+++ src/Resources/config/compatibility/services_6.3.5.0.xml
@@ -6,10 +6,11 @@
 
     <services>
 
-        <service id="Kiener\MolliePayments\Compatibility\Storefront\Route\PaymentMethodRoute\Voucher\HideVoucherPaymentMethodRoute63"
+        <service id="Kiener\MolliePayments\Compatibility\Storefront\Route\PaymentMethodRoute\RemovePaymentMethodRoute63"
                  decorates="Shopware\Core\Checkout\Payment\SalesChannel\PaymentMethodRoute">
-            <argument type="service" id="Kiener\MolliePayments\Compatibility\Storefront\Route\PaymentMethodRoute\Voucher\HideVoucherPaymentMethodRoute63.inner"/>
-            <argument type="service" id="Kiener\MolliePayments\Components\Voucher\Service\VoucherRemover"/>
+            <argument type="service" id="Kiener\MolliePayments\Compatibility\Storefront\Route\PaymentMethodRoute\RemovePaymentMethodRoute63.inner"/>
+
+            <argument type="tagged" tag="mollie_payments.payment_method.remover.sw63"/>
         </service>
 
         <service id="mollie_payments.mail_service" class="Kiener\MolliePayments\Service\Mail\MailService63">
@@ -17,12 +18,6 @@
             <argument type="service" id="Shopware\Core\Content\MailTemplate\Service\MessageFactory"/>
             <argument type="service" id="Shopware\Core\Content\MailTemplate\Service\MailSender"/>
             <argument type="service" id="mollie_payments.logger"/>
-        </service>
-
-        <service id="Kiener\MolliePayments\Compatibility\Storefront\Route\PaymentMethodRoute\Subscription\SubscriptionPaymentMethodRoute63"
-                 decorates="Shopware\Core\Checkout\Payment\SalesChannel\PaymentMethodRoute">
-            <argument type="service" id="Kiener\MolliePayments\Compatibility\Storefront\Route\PaymentMethodRoute\Voucher\HideVoucherPaymentMethodRoute63.inner"/>
-            <argument type="service" id="Kiener\MolliePayments\Components\Subscription\Services\PaymentMethodRemover\SubscriptionRemover"/>
         </service>
 
     </services>
