+++ src/Resources/config/compatibility/services_6.4.xml
@@ -6,22 +6,11 @@
 
     <services>
 
-        <service id="Kiener\MolliePayments\Compatibility\Storefront\Route\PaymentMethodRoute\Voucher\HideVoucherPaymentMethodRoute64"
+        <service id="Kiener\MolliePayments\Compatibility\Storefront\Route\PaymentMethodRoute\RemovePaymentMethodRoute64"
                  decorates="Shopware\Core\Checkout\Payment\SalesChannel\PaymentMethodRoute">
-            <argument type="service" id="Kiener\MolliePayments\Compatibility\Storefront\Route\PaymentMethodRoute\Voucher\HideVoucherPaymentMethodRoute64.inner"/>
-            <argument type="service" id="Kiener\MolliePayments\Components\Voucher\Service\VoucherRemover"/>
-        </service>
+            <argument type="service" id="Kiener\MolliePayments\Compatibility\Storefront\Route\PaymentMethodRoute\RemovePaymentMethodRoute64.inner"/>
 
-        <service id="Kiener\MolliePayments\Compatibility\Storefront\Route\PaymentMethodRoute\MollieLimits\MollieLimitsPaymentMethodRoute64"
-                 decorates="Shopware\Core\Checkout\Payment\SalesChannel\PaymentMethodRoute">
-            <argument type="service" id="Kiener\MolliePayments\Compatibility\Storefront\Route\PaymentMethodRoute\MollieLimits\MollieLimitsPaymentMethodRoute64.inner"/>
-            <argument type="service" id="Kiener\MolliePayments\Components\MollieLimits\Service\MollieLimitsRemover"/>
-        </service>
-
-        <service id="Kiener\MolliePayments\Compatibility\Storefront\Route\PaymentMethodRoute\Subscription\SubscriptionPaymentMethodRoute64"
-                 decorates="Shopware\Core\Checkout\Payment\SalesChannel\PaymentMethodRoute">
-            <argument type="service" id="Kiener\MolliePayments\Compatibility\Storefront\Route\PaymentMethodRoute\Subscription\SubscriptionPaymentMethodRoute64.inner"/>
-            <argument type="service" id="Kiener\MolliePayments\Components\Subscription\Services\PaymentMethodRemover\SubscriptionRemover"/>
+            <argument type="tagged" tag="mollie_payments.payment_method.remover.sw64"/>
         </service>
 
         <service id="Kiener\MolliePayments\Compatibility\Storefront\Route\PaymentMethodRoute\Cache\CachedPaymentMethodRoute64">
